Extract the BigDecimal and java.time encoders/decoders to a new scala object
